requirements.txt
streamlit
google-cloud-bigquery
google-cloud-bigquery-storage